requirements.txt
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
tensorflow>=2.15   # ili tensorflow>=2.15 ako želiš “punu” verziju
